Create structure,union and enumeration
shape object encompassing both Square and Rectangle, you can use a structure in C to represent each shape. The Shape structure can then use a union to allow it to store either a Square or a Rectangle, as well as an enumeration to indicate which type of shape is currently being used.